Security hardening: authenticated handshake, TTL, replay protection, v2 session
Closes 12 audit findings spanning identity spoofing, DoS, and replay vectors.
Bumps wire protocol to v2 (NRN1 magic + lowercase session magics) — old peers
fail loudly at parse time rather than silently mis-interpreting.

Transport
- Authenticated TCP handshake with per-side nonces and Ed25519 challenge-response
  binds the claimed pub_key to the live connection. Closes LAN identity spoofing.
- 256-permit semaphore + 10s timeout caps unauthenticated handshakes in flight.

Router
- TTL via watermark hop counter (cap 32) prevents forwarding loops from
  disagreeing cuckoo state.
- lookup_by_tag_excluding(from) breaks 2-cycles caused by bidirectional cuckoo
  gossip (peer.cuckoo carries our own tag back to us).
- forward_sem caps in-flight tokio::spawn forwards at 4096.
- fwd.from = self.pub_key on forward; previously the original source's pub_key
  leaked at every hop, defeating source-privacy.
- coord_table capped at 16 384 entries; NaN/Inf coords rejected.
- pad_payload guards against u16 length wrap; mtu() lowered to 65 000.
- Cuckoo merge is now a true set-union with kick cascade, not position-wise OR
  (which silently dropped colliding fingerprints).

Session (v2 wire format)
- SessionInit/Ack now carry timestamp_ms and recipient_ed_pub. Signature binds
  to intended responder and a ±60s window, blocking cross-target and stale
  replays.
- handle_ack rejects unsolicited ACKs (no pending initiate).

Admin
- Atomic 0o600 socket: tighten umask before bind + explicit chmod after, fail
  loudly if either step fails. Closes TOCTOU window allowing local addPeer SSRF.

Config
- Refuse world-readable config files (any non-zero group/other bits).
- Ephemeral keys require NORN_ALLOW_EPHEMERAL_KEY=1; default policy refuses.
- genconfig writes with O_CREAT|O_EXCL|0o600 atomically.
- zeroize intermediate key bytes; SigningKey self-zeroizes on drop.

Production-grade docs and CI
- SECURITY.md: threat model, claimed properties, known weaknesses, disclosure
  policy, supported-version matrix.
- PROTOCOL.md: normative byte-for-byte wire-format spec for v2.
- deny.toml: license allow-list, ban openssl-sys/rust-crypto, deny unknown
  registries.
- ci.yml: stable+beta matrix, clippy -D warnings, cargo-audit, cargo-deny,
  fuzz smoke (60s per target).

Tests
- 271 unit + 11 integration + 3 binary tests all pass; clippy clean with
  -D warnings; flaky four_node_linear_chain now stable 60/60.

+++ src/tun.rs
@@ -156,7 +156,7 @@ pub async fn start(
 
     // TUN MTU: norn payload minus session overhead (ChaCha20 tag=16, length=2,
     // padding up to 255 bytes worst-case). 65200 fits comfortably.
-    let tun_mtu = (conn.mtu() as u64).saturating_sub(300).max(1280) as u16;
+    let tun_mtu = conn.mtu().saturating_sub(300).max(1280) as u16;
 
     // Create TUN device
     let mut tun_config = tun2::Configuration::default();
